bubble-cyclegan/requirements.txt
tensorboard
torchinfo
albumentations
